server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
google-genai>=1.5.0
google-cloud-texttospeech>=2.16.0
pydantic>=2.9.0
python-dotenv>=1.0.0
